# server/openapi.yaml
openapi: 3.0.3
info:
  title: Stablecog API
  description: API Specification for stablecog
  version: "2"
servers:
  - url: http://localhost:13337
paths:
...
  /v1/gallery:
    get:
      description: "Gallery: Search"
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            format: optional, default:50
          example: 1
        - name: next
          in: query
          schema:
            type: integer
            format: optional
          example: 2
        - name: search
          in: query
          schema:
            type: string
            format: optional
          example: A portrait of a cat by Van Gogh
      responses:
        "200":
          description: JSON object with gallery hits and metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse400"
        "401":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse401"
...
  /v1/user/gallery:
    put:
      description: "Gallery: Submit (User-only)"
      requestBody:
        description: JSON Object with generation output IDs to submit
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitGalleryRequest"
      responses:
        "200":
          description: JSON Object indicating how many items were submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmittedResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse400"
        "401":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse401"
  /v1/admin/gallery:
    put:
      description: "Gallery: Approve/Reject (Admin)"
      security:
        - supabase: []
      requestBody:
        description: JSON Object with review action and output IDs
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewGalleryRequest"
      responses:
        "200":
          description: Number of items approved/rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatedResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse400"
        "401":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse401"
...
components:
  securitySchemes:
    supabase:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ErrorResponse400:
      type: object
      properties:
        error:
          type: string
          format: insufficient_credits, json_parse_error, invalid_stream_id, invalid_model_id, invalid_scheduler_id, invalid_image_url, invalid_output_id, image_url_width_height_error, output_not_found, image_already_upscaled
    ErrorResponse401:
      type: object
      properties:
        error:
          type: string
          format: unauthorized
...
    SubmitGalleryRequest:
      type: object
      properties:
        generation_output_ids:
          type: array
          items:
            type: string
            format: uuidv4
    SubmittedResponse:
      type: object
      properties:
        submitted:
          type: integer
      description: The number of items successfully submitted to gallery
...
    ReviewGalleryRequest:
      type: object
      properties:
        action:
          type: string
          format: approve, reject
        generation_output_ids:
          type: array
          items:
            type: string
            format: uuidv4
    UpdatedResponse:
      type: object
      properties:
        updated:
          type: integer
      description: The number of items successfully updated
...
    GalleryResponse:
      type: object
      properties:
        page:
          type: integer
        next:
          type: integer
        hits:
          type: array
          items:
            type: object
            properties:
              image_url:
                type: string
                format: URL
              upscaled_image_url:
                type: string
                format: URL, optional
              width:
                type: integer
                format: int32
              height:
                type: integer
                format: int32
              inference_steps:
                type: integer
                format: int32
              guidance_scale:
                type: number
                format: float32
              seed:
                type: integer
              model_id:
                type: string
                format: uuidv4
              scheduler_id:
                type: string
                format: uuidv4
              prompt_text:
                type: string
              prompt_id:
                type: string
                format: uuidv4
              negative_prompt_text:
                type: string
                format: optional
              negative_prompt_id:
                type: string
                format: uuidv4
              created_at:
                type: string
                format: isoTimeString
              updated_at:
                type: string
                format: isoTimeString
...